cms_form: refactor tests w/ odoo-test-helper
