+++ cms_form/tests/common.py
@@ -2,16 +2,11 @@
 # License LGPL-3.0 or later (http://www.gnu.org/licenses/lgpl.html).
 
 from lxml import html
+from odoo_test_helper import FakeModelLoader
 
 from odoo.tests.common import HttpCase, SavepointCase
 
-from .utils import (
-    fake_request,
-    fake_session,
-    session_store,
-    setup_test_model,
-    teardown_test_model,
-)
+from .utils import fake_request, fake_session, session_store
 
 
 def get_form(env, form_model, req=None, session=None, ctx=None, sudo_uid=None, **kw):
@@ -38,20 +33,32 @@ def get_form(env, form_model, req=None, session=None, ctx=None, sudo_uid=None, *
 class FakeModelMixin(object):
     """Mixin to setup fake models just for testing."""
 
-    # override this in your test case to inject new models on the fly
-    TEST_MODELS_KLASSES = []
+    @staticmethod
+    def _get_test_models():
+        return ()
 
-    @classmethod
-    def _setup_models(cls):
+    @staticmethod
+    def _setup_models(class_or_instance):
         """Setup new fake models for testing."""
-        for kls in cls.TEST_MODELS_KLASSES:
-            setup_test_model(cls.env, kls)
+        if hasattr(class_or_instance, "loader"):
+            # Especially for HttpCase: try to setup models only once
+            return
+        class_or_instance.loader = FakeModelLoader(
+            class_or_instance.env, class_or_instance.__module__
+        )
+        class_or_instance.loader.backup_registry()
+        test_models = class_or_instance._get_test_models()
+        if test_models:
+            class_or_instance.loader.update_registry(test_models)
+            for klass in test_models:
+                # Make them available on the test class
+                setattr(class_or_instance, klass.__name__, klass)
 
-    @classmethod
-    def _teardown_models(cls):
+    @staticmethod
+    def _teardown_models(class_or_instance):
         """Wipe fake models once tests have finished."""
-        for kls in cls.TEST_MODELS_KLASSES:
-            teardown_test_model(cls.env, kls)
+        if hasattr(class_or_instance, "loader"):
+            class_or_instance.loader.restore_registry()
 
 
 class HTMLRenderMixin(object):
@@ -77,6 +84,16 @@ class FormTestCase(SavepointCase, FakeModelMixin):
 
     at_install = False
     post_install = True
+
+    @classmethod
+    def setUpClass(cls):
+        super().setUpClass()
+        cls._setup_models(cls)
+
+    @classmethod
+    def tearDownClass(cls):
+        cls._teardown_models(cls)
+        super().tearDownClass()
 
     def get_form(self, form_model, **kw):
         return get_form(self.env, form_model, **kw)
@@ -104,13 +121,11 @@ class FormHttpTestCase(HttpCase, FakeModelMixin, HTMLRenderMixin):
     def setUp(self):
         # HttpCase has no ENV on setUpClass we have to setup fake models here
         super().setUp()
-        for kls in self.TEST_MODELS_KLASSES:
-            setup_test_model(self.env, kls)
+        self._setup_models(self)
 
     def tearDown(self):
         # HttpCase has no ENV on setUpClass
-        for kls in self.TEST_MODELS_KLASSES:
-            teardown_test_model(self.env, kls)
+        self._teardown_models(self)
         super().tearDown()
 
     def html_get(self, url):
